open link from markdown into external browser
